更新刪除刊登單後的動作(刪除刊登單,刪除訂單,刪除預約單,刪除合約,刪除其他付款)以上皆為update delete_time

+++ src/main/java/controller/PublishListController.java
@@ -15,21 +15,20 @@ import javax.servlet.http.HttpServletResponse;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseOptions;
-import com.google.firebase.messaging.FirebaseMessaging;
-import com.google.firebase.messaging.FirebaseMessagingException;
-import com.google.firebase.messaging.Message;
-import com.google.firebase.messaging.Notification;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 
+import member.bean.Agreement;
 import member.bean.Appointment;
 import member.bean.Order;
 import member.bean.Publish;
+import service.AgreementService;
 import service.AppointmentService;
 import service.CityService;
 import service.MemberService;
 import service.NotificationService;
 import service.OrderService;
+import service.OtherPayService;
 import service.PublishService;
 
 /**
@@ -82,31 +81,90 @@ public class PublishListController extends HttpServlet {
 			} else if (clientReq.get("action").getAsString().equals("pubishDelete")) {
 				// 被刪除的刊登單Id
 				int publishId = clientReq.get("publishId").getAsInt();
-				int notified=publishService.selectById(publishId).getOwnerId();
+				int notified = publishService.selectById(publishId).getOwnerId();
 				JsonObject resp = new JsonObject();
 				if (publishService.deleteById(publishId) == 1) {
 					// 刪除預約單的通知及預約單
 					List<Appointment> appointments = new AppointmentService().selectAppointmentByPublishId(publishId);
-					if (appointments.size() > 0) {
+					if (appointments.size() > 0&& appointments!=null) {
+						int count1 = 0;
 						for (Appointment appointment : appointments) {
+							// 刪除預約
 							appointmentService.deleteById(appointment.getAppointmentId());// 加Delete_Time
-							notificationService.deleteAppointment(notified,appointment.getAppointmentId());
+							// 房東
+							count1 = notificationService.deleteAppointment(appointment.getOwnerId(),
+									appointment.getAppointmentId());
+
+							// 刪除及即時更新通知房客
+							if (notificationService.deleteAppointment(appointment.getTenantId(),
+									appointment.getAppointmentId()) > 0) {
+								String memberToken2 = new MemberService().selectById(appointment.getTenantId())
+										.getToken();
+								if (memberToken2 != null) {
+									NotificationController.sendSingleFcmNoNotification(memberToken2);
+								}
+							}
+						}
+						// 如果房東通知有刪除
+						if (count1 > 0) {
+							// 房東僅需觸發一次
+							String memberToken1 = new MemberService().selectById(notified).getToken();
+							if (memberToken1 != null) {
+								NotificationController.sendSingleFcmNoNotification(memberToken1);
+
+							}
 						}
 					}
+
 					// 刪除訂單的通知及訂單
 					List<Order> orders = orderService.selectAllByPublishID(publishId);
-					if (orders.size() > 0) {
+					List<Agreement> agreements=new ArrayList<Agreement>();
+					AgreementService agreementService=new AgreementService();
+					if (orders.size() > 0 && orders!=null) {
+						int count1 = 0;
 						for (Order order : orders) {
-							orderService.deleteById(order.getOrderId());// 加Delete_Time
-							notificationService.deleteOrder(notified, order.getOrderId()); 
+							//取得跟此訂單有關的所有合約(刪除其他付款用)
+							if(agreementService.selectByOrderId(order.getOrderId())!=null) {
+								agreements.add(agreementService.selectByOrderId(order.getOrderId()));
+							}
+							// 刪除合約
+							agreementService.deleteByOrderId(order.getOrderId());
+							// 刪除訂單
+							orderService.deleteByPublishId(order.getPublishId());// 加Delete_Time
+							// 刪除訂單通知
+							count1 = notificationService.deleteOrder(notified, order.getOrderId());
+							// 刪除及即時更新通知房客
+							if (notificationService.deleteOrder(order.getTenantId(), order.getOrderId()) > 0) {
+								String memberToken2 = new MemberService().selectById(order.getTenantId()).getToken();
+								if (memberToken2 != null) {
+									NotificationController.sendSingleFcmNoNotification(memberToken2);
+								}
+							}
+						}
+						//如果房東通知有刪除
+						if (count1 > 0) {
+							// 房東僅需觸發一次
+							String memberToken1 = new MemberService().selectById(notified).getToken();
+							if (memberToken1 != null) {
+								NotificationController.sendSingleFcmNoNotification(memberToken1);
+							}
 						}
 					}
-					// 觸發通知
-					String memberToken = new MemberService().selectById(appointments.get(0).getOwnerId()).getToken();
-					if (memberToken != null) {	
-						NotificationController.sendSingleFcmNoNotification(memberToken);
+					//刪除其他付款
+					if(agreements.size()>0 && agreements!=null) {
+						System.out.println(agreements.size()+"");
+						OtherPayService otherPayService=new OtherPayService();
+//						int count=0;//測試用
+						for(Agreement agreement:agreements) {
+							otherPayService.deleteByAgreementId(agreement.getAgreementId());
+//							count+=otherPayService.deleteByAgreementId(agreement.getAgreementId()); //測試用
+							
+						}
+//						System.out.println(count+""); //測試用
 					}
 					resp.addProperty("result", true);
+				
+					//刪除其他付款
 				} else {
 					resp.addProperty("result", false);
 				}
@@ -131,6 +189,5 @@ public class PublishListController extends HttpServlet {
 			}
 		}
 	}
-
 
 }
